refactor location to /a32nx dir
